feat: move distance calculation to common and add test
