+++ packages/common/index.ts
@@ -5,3 +5,4 @@ export * as DatabaseTypes from "./src/databaseTypes/types";
 export * from "./src/databaseTypes/CollectionNames";
 export * from "./src/AppLinks";
 export * from "./src/GlobalParams";
+export * from "./src/DistanceHelper";
