+++ packages/common/package.json
@@ -5,9 +5,23 @@
   "license": "MIT",
   "private": true,
   "devDependencies": {
+    "@types/jest": "^29.5.12",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.3",
     "typescript": "*"
   },
   "dependencies": {
     "moment-timezone": "^0.5.48"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "jest": {
+    "testEnvironment": "jest-environment-node",
+    "transform": {
+      "^.+\\.ts$": [
+        "ts-jest"
+      ]
+    }
   }
 }
